refactor(starterkit): split the new converters into one type per file
- Fold `StringWhiteSpaceToBoolConverter` into `StringEmptyToBoolConverter`
  behind a `StringEmptiness` setting, so one converter answers every flavour
  of "is this string absent?" — `NullOrEmpty` stays the zero value, so an
  already authored converter keeps its behaviour.
- Keep every setting enum top-level: nesting one in a generic converter makes
  `EnumToBoolConverter<A>.EnumMatch` and `<B>.EnumMatch` different types, which
  no caller can spell and no test can share.
- Document the shipped catalogue — composition, failure modes, converter
  assets, and the enum and selection families.
- Drop the Enum-to-OptionData entry from `KnownGaps`:
  `EnumToDropdownOptionDataConverter` already fills it.
