+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Bools/BoolLogicConverter.cs
@@ -5,30 +5,6 @@ using UnityEngine;
 namespace Aspid.MVVM.StarterKit
 {
     /// <summary>
-    /// The boolean operations <see cref="BoolLogicConverter"/> can apply.
-    /// </summary>
-    public enum LogicOperation
-    {
-        /// <summary>Both the bound value and the operand must be <see langword="true"/>.</summary>
-        And,
-
-        /// <summary>Either the bound value or the operand must be <see langword="true"/>.</summary>
-        Or,
-
-        /// <summary>Exactly one of the bound value and the operand must be <see langword="true"/>.</summary>
-        Xor,
-
-        /// <summary>The negation of <see cref="And"/>.</summary>
-        Nand,
-
-        /// <summary>The negation of <see cref="Or"/>.</summary>
-        Nor,
-
-        /// <summary>The negation of <see cref="Xor"/> — the two must agree.</summary>
-        Xnor,
-    }
-
-    /// <summary>
     /// Combines a bound boolean with an authored one.
     /// </summary>
     /// <remarks>
@@ -50,8 +26,8 @@ namespace Aspid.MVVM.StarterKit
         /// <param name="operand">The authored value the bound one combines with.</param>
         public BoolLogicConverter(LogicOperation operation, bool operand)
         {
-            _operation = operation;
             _operand = operand;
+            _operation = operation;
         }
 
         /// <summary>
