+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Bools/LogicOperation.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 7330a6ef4da8424aa5597d6fe8ef2399
+timeCreated: 1786963263
