+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Bools/StringEmptiness.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 20dd539ffed54a449ba527f1f993c513
+timeCreated: 1787000000
