+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Bools/StringEmptyToBoolConverter.cs
@@ -1,40 +1,55 @@
 using System;
+using UnityEngine;
 
 // ReSharper disable once CheckNamespace
 namespace Aspid.MVVM.StarterKit
 {
     /// <summary>
-    /// Converts string values to boolean based on empty check, with optional inversion.
+    /// Converts a string to a boolean based on whether it is absent, with optional inversion.
     /// </summary>
+    /// <remarks>
+    /// <see cref="StringEmptiness.NullOrWhiteSpace"/> is what "did the user type anything?" usually
+    /// means, since a string of spaces is not empty but reads as one.
+    /// </remarks>
     [Serializable]
     public class StringEmptyToBoolConverter : IConverterStringToBool
     {
-        [UnityEngine.SerializeField]
-        private bool _isInvert;
+        [Tooltip("What counts as an absent string.")]
+        [SerializeField] private StringEmptiness _emptiness = StringEmptiness.NullOrEmpty;
 
-        /// <summary>
-        /// Initializes a new instance of the <see cref="StringEmptyToBoolConverter"/> class.
-        /// </summary>
-        public StringEmptyToBoolConverter()
-            : this(isInvert: false) { }
+        [Tooltip("Invert the result — true when the string has content.")]
+        [SerializeField] private bool _isInvert;
 
-        /// <summary>
-        /// Initializes a new instance of the <see cref="StringEmptyToBoolConverter"/> class.
-        /// </summary>
-        /// <param name="isInvert">If <c>true</c>, inverts the result of the empty check. Default is <c>false</c>.</param>
+        public StringEmptyToBoolConverter() { }
+
+        /// <param name="isInvert">When <see langword="true"/>, inverts the result.</param>
         public StringEmptyToBoolConverter(bool isInvert)
+            : this(StringEmptiness.NullOrEmpty, isInvert) { }
+
+        /// <param name="emptiness">What counts as an absent string.</param>
+        /// <param name="isInvert">When <see langword="true"/>, inverts the result.</param>
+        public StringEmptyToBoolConverter(StringEmptiness emptiness, bool isInvert = false)
         {
+            _emptiness = emptiness;
             _isInvert = isInvert;
         }
 
         /// <summary>
-        /// Converts a string to boolean based on whether it is null or empty.
+        /// Tests whether the specified string is absent under the configured <see cref="StringEmptiness"/>.
         /// </summary>
-        /// <param name="value">The string to check.</param>
-        /// <returns><c>true</c> if the value is null or empty (or not if inverted), otherwise <c>false</c>.</returns>
+        /// <param name="value">The string to test.</param>
+        /// <returns><see langword="true"/> when the string is absent, inverted when configured.</returns>
+        /// <exception cref="ArgumentOutOfRangeException">Thrown when the emptiness mode is not a declared value.</exception>
         public bool Convert(string? value)
         {
-            var isEmpty = string.IsNullOrEmpty(value);
+            var isEmpty = _emptiness switch
+            {
+                StringEmptiness.Null => value is null,
+                StringEmptiness.NullOrEmpty => string.IsNullOrEmpty(value),
+                StringEmptiness.NullOrWhiteSpace => string.IsNullOrWhiteSpace(value),
+                _ => throw new ArgumentOutOfRangeException(nameof(_emptiness), _emptiness, null)
+            };
+
             return _isInvert ? !isEmpty : isEmpty;
         }
     }
