+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Bools/StringMatch.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 18dbba4093c640bea8d7d88fa67e7dd2
+timeCreated: 1787000007
