+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Bools/StringMatchToBoolConverter.cs
@@ -5,24 +5,6 @@ using UnityEngine;
 namespace Aspid.MVVM.StarterKit
 {
     /// <summary>
-    /// How <see cref="StringMatchToBoolConverter"/> compares a bound string with the authored one.
-    /// </summary>
-    public enum StringMatch
-    {
-        /// <summary>The whole string must match.</summary>
-        Equals,
-
-        /// <summary>The string must contain the authored text.</summary>
-        Contains,
-
-        /// <summary>The string must begin with the authored text.</summary>
-        StartsWith,
-
-        /// <summary>The string must end with the authored text.</summary>
-        EndsWith,
-    }
-
-    /// <summary>
     /// Tests a bound string against an authored one.
     /// </summary>
     /// <remarks>
@@ -57,10 +39,10 @@ namespace Aspid.MVVM.StarterKit
             bool ignoreCase = true,
             bool isInvert = false)
         {
-            _match = match;
             _text = text;
-            _ignoreCase = ignoreCase;
+            _match = match;
             _isInvert = isInvert;
+            _ignoreCase = ignoreCase;
         }
 
         /// <summary>
@@ -71,51 +53,20 @@ namespace Aspid.MVVM.StarterKit
         /// <exception cref="ArgumentOutOfRangeException">Thrown when the match mode is not a declared value.</exception>
         public bool Convert(string? value)
         {
+            // A converter built in code can be handed a null text; Unity only ever writes an empty one.
+            var text = _text ?? string.Empty;
             var comparison = _ignoreCase ? StringComparison.OrdinalIgnoreCase : StringComparison.Ordinal;
 
             var matched = value is not null && _match switch
             {
-                StringMatch.Equals => value.Equals(_text, comparison),
-                StringMatch.Contains => value.IndexOf(_text ?? string.Empty, comparison) >= 0,
-                StringMatch.StartsWith => value.StartsWith(_text ?? string.Empty, comparison),
-                StringMatch.EndsWith => value.EndsWith(_text ?? string.Empty, comparison),
+                StringMatch.Equals => value.Equals(text, comparison),
+                StringMatch.Contains => value.IndexOf(text, comparison) >= 0,
+                StringMatch.StartsWith => value.StartsWith(text, comparison),
+                StringMatch.EndsWith => value.EndsWith(text, comparison),
                 _ => throw new ArgumentOutOfRangeException(nameof(_match), _match, null)
             };
 
             return _isInvert ? !matched : matched;
-        }
-    }
-
-    /// <summary>
-    /// Converts a string to a boolean based on whether it is null, empty or whitespace.
-    /// </summary>
-    /// <remarks>
-    /// <see cref="StringEmptyToBoolConverter"/> asks whether a string is empty; this one also counts
-    /// a string of spaces as empty, which is what "did the user type anything?" usually means.
-    /// </remarks>
-    [Serializable]
-    public sealed class StringWhiteSpaceToBoolConverter : IConverterStringToBool
-    {
-        [Tooltip("Invert the result — true when the string has content.")]
-        [SerializeField] private bool _isInvert;
-
-        public StringWhiteSpaceToBoolConverter() { }
-
-        /// <param name="isInvert">If <see langword="true"/>, inverts the result.</param>
-        public StringWhiteSpaceToBoolConverter(bool isInvert)
-        {
-            _isInvert = isInvert;
-        }
-
-        /// <summary>
-        /// Tests whether the specified string is null, empty or whitespace.
-        /// </summary>
-        /// <param name="value">The string to test.</param>
-        /// <returns><see langword="true"/> when the string has no content, inverted when configured.</returns>
-        public bool Convert(string? value)
-        {
-            var isBlank = string.IsNullOrWhiteSpace(value);
-            return _isInvert ? !isBlank : isBlank;
         }
     }
 }
