+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Bools/StringMatchToBoolConverter.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: eed35eaea952404a88a1ad41b45681cf
+timeCreated: 1786963426
