--- Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Bools/StringToBoolConverters.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: b33de229b1fdf403e9c1e9b3f9737c6e
