+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Enums/EnumNameSource.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: bae4a1fbbf6448c3a3a5a99b54dcbce7
+timeCreated: 1787000021
