+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Enums/EnumToBoolConverter.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 8ad1c62f8a684fb1a0dead0f2d46d895
+timeCreated: 1786966751
