+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Enums/EnumToStringConverter.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: a2aa7a6885af4a798df7bd95689e40b2
+timeCreated: 1786966790
