+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Enums/EnumToValueConverter.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 3963ef375a544bd09a994e4c55444595
+timeCreated: 1786966685
