+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Runtime/Converters/Objects/IndexMode.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 7c62be97babf49be9c09a6e958ece5b6
+timeCreated: 1787000028
